refactor: move getParentKey into utils
