+++ src/test/test_job.ts
@@ -4,9 +4,8 @@
 import { Job, Queue, QueueScheduler } from '../classes';
 import { QueueEvents } from '../classes/queue-events';
 import { Worker } from '../classes/worker';
-import { getParentKey } from '../classes/flow-producer';
 import { JobsOptions } from '../interfaces';
-import { delay, removeAllQueueData } from '../utils';
+import { delay, getParentKey, removeAllQueueData } from '../utils';
 import * as chai from 'chai';
 import * as chaiAsPromised from 'chai-as-promised';
 import * as IORedis from 'ioredis';
